// src/components/ui/burger-constructor/type.ts
import {
  TOrder,
  TIngredient,
  TConstructorIngredient
} from '../../../utils/types';

export type BurgerConstructorUIProps = {
  constructorItems: {
    bun: TIngredient | null;
    ingredients: TConstructorIngredient[];
  };
  orderRequest: boolean;
  price: number;
  orderModalData: TOrder | null;
  onOrderClick: () => void;
  closeOrderModal: () => void;
};